Add auto-repair for crashing etcd in operator (#806)
Summary: We've noticed that a lot of clusters running etcd are in a
state where etcd is crashing. This is very likely to happen if etcd
loses quorum. This PR updates the operator to detect that state and
attempt to fix it by restarting etcd.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Skaffold deploy operator

---------

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>