Improve Command Palette and Breadcrumb sorting
Summary: There were still a few scenarios where the filtering was a bit too aggressive, or the sorting of results didn't make much sense. This should fix quite a few of them.

Relevant Issues:

Test Plan: Try all of the following searches in the `script` breadcrumb, and a few of them in the Command Palette as well.
They should both get similar results:

- The search `exec` should now find `bpftrace/EXEC_snoop` before it finds weaker matches
- The search `pod` should now greatly prefer scripts that have the string `pod` in them, even over shorter-named scripts that have vaguely similar spelling.
- The search `px/ace' should now do the same: `PX/namespACEs` should now be a stronger match than `PX/nodE`.
- And the same for `px/stat`.
- Finally, most searches should keep more results than before (further past "definitely relevant" into "eh, it might be what you meant" territory), but the most relevant ones should still come up first.

Type of change: /kind bug

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>