Fix off by one error in `amqp/parse.cc`. (#946)
Summary: We fix an off-by-one error in `amqp/parse.cc` that was causing
segfaults when parsing traffic for the same.

We also change the parse state to `kNeedsMoreData` (vs. `kInvalid` its
previous state update) if the parser returns early when the buffer does
not have enough data to read the entire payload.

Closes #942.

We opened #945 to track test coverage.

Type of change: /kind bug

Test Plan: Ran `stirling_wrapper` both with and without the fix on a
node with `px-sock-shop` deployed (with load testing up). Without the
off-by-one fix, we wbserved segfaults matching the original issue
reported. With the fix, observed no segfaults.

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>