[perf_tool] Add thin wrapper around bigquery Table, to store the schema alongside it. (#927)
Summary: The `bigquery.Schema` is required for some bigquery operations,
so its useful to be able to access the `bigquery.Table` and
`bigquery.Schema` in a unified way. This diff adds a thin wrapper around
`bigquery.Table` that also stores the schema. In addition, this diff
adds a unified way to create a bigquery Table/Schema pair (creating the
gcloud resources if necessary). Currently, this is only used in
`perf_tool` but we could consider pulling it up for our other services
that rely on bigquery.

Type of change: /kind cleanup

Test Plan: Tested the table (and dataset) creation with the `perf_tool
run` command coming in future diffs.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>