Add target_rps parameter to protocol_loadtest (#926)
Summary: Add `target_rps` parameter to `protocol_loadtest`. This
parameter sets a target for requests per second for the loadtest. It is
only a limit on the RPS, so if the target RPS can't be achieved given
the CPU/network of the cluster, the RPS will end up lower than the
target.
This is useful to ensure consistency of the `protocol_loadtest` across
different CPU loads. For example, without this change running the
`protocol_loadtest` on an empty cluster would reach significantly higher
RPS than running on a loaded cluster.
Also fixes a misnaming of one of the return values in the httpclient
(`resp_size` -> `body_size`).

Type of change: /kind test-infra

Test Plan: Tested a skaffold deployment of the `protocol_loadtest`, saw
that it reached the target RPS (with roughly 10% margin).

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>