[perf_tool] Add `cluster` package interface (#928)
Summary: `cluster.Provider` specifies the interface for any providers of
k8s clusters. This interface returns a `cluster.Context` pointing to a
k8s cluster that matches the spec provided to
`cluster.Provider.GetCluster`. The `cluster.Context` stores the
kubeconfig, and allows consumers to either access the k8s golang API
client connected to the given cluster or run commands with the given
cluster using the `KUBECONFIG` environment variable. There are currently
no implementations of `cluster.Provider` but future diffs will add a
`local` provider for using your local kubeconfig, and a `gke` provider
for creating clusters on GKE.

Type of change: /kind test-infra

Test Plan: This just adds the interface, the implementations of the
interface are tested in future diffs.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>