[perf_tool] Add BatchInserter for bigquery (#961)
Summary: BigQuery `Inserter().Put` calls can be quite slow, so we don't
want to run that for every single metric row we receive during an
experiment.
This adds a utility that reads from a channel of arbitrary data, and
batches it up before sending it to bigquery. This reduces the number of
`Put` calls.

Type of change: /kind test-infra

Test Plan: Tested that rows sent to the BatchInserter show up in
bigquery. Also tested that if not enough rows are sent the timeout will
trigger a push anyway.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>