[perf_tool] Add `tags` to ExperimentSpec (#973)
Summary: Adds a `tags` field to the `ExperimentSpec`. This field can be
used for tagging an experiment with useful information. For example,
`workload/http-loadtest` to mark that the experiment ran the
`http-loadtest` workload, or `ignore` to mark that the experiment should
be ignored in aggregate datastudio views.

Type of change: /kind test-infra

Test Plan: No functional changes, just updates the protobuf definition.
Tested with future PRs that the tags are accessible in bigquery.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>