[perf_tool/deploy] Allow patching rendered YAML before deploy (#977)
Summary: Allows patching rendered workload YAMLs via kustomize. This
will be used by both skaffold, and prerendered deploy steps. This allows
workloads to have parameters that can be easily changed in the spec
without having to mess with the prerendered yamls, or the skaffold
config.

Type of change: /kind test-infra

Test Plan: Tested as part of broader changes, that the patching works as
expected.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>