[perf_tool/runner] Add sleeps (#1000)
Summary: Add sleeps based on the `RunSpec`. We sleep once after `vizier`
is deployed and before the other workloads are deployed, as a burn-in
period. We then sleep after the rest of the workloads are deployed for
the duration of the rest of the experiment.

Type of change: /kind test-infra

Test Plan: Tested with the full `Runner` implementation.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>