Always scroll hovered item into view in Command Palette
Summary: Without this change, the browser already tries to scroll the selected option into view. However, due to `position: sticky`, there's a header that can obscure it. To account for that header, we use `scroll-margin-top` and force the browser to try again with `scrollIntoView`, which accounts for that property.
Fixes #1045

Relevant Issues: #1045

Test Plan: Open the Command Palette. Enter a query like script:. Using the arrow keys, scroll down a couple of pages, then scroll back up the same way. When moving to an option that's above the current scroll view, it should now fully scroll that option into view.

Type of change: /kind bug

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>