Command Palette: Fix missing quote logic
Summary: When opening the Command Palette while working on the Scratch Pad script, it was failing to quote the space in the name (resulting in an invalid command).
This makes it use an existing utility function to correctly quote both the script name and any arguments that need it, when opening the Command Palette in the live view.

Test Plan: Open the Command Palette when editing the Scratch Pad script or a script whose arg values include a character that requires quoting.

Type of change: /kind bug

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>