Add securityContext for vizier and vizier deps (#1040)
Summary: This pull request aims to enhance the security of the vizier
and vizier dependencies by adding securityContext to the Kubernetes
workloads and removing redundant permissions from containers that will
be deployed on end users' Kubernetes clusters. This matters because
security is a crucial aspect of any project, especially when dealing
with sensitive data or running in a production environment. By
implementing security best practices such as least privilege access and
securityContext, we can reduce the attack surface and minimize the
impact of potential security breaches.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Build vizier and vizier deps Kubernetes manifest files and
deploy them into a cluster.

Signed-off-by: Isala Piyarisi <isala@wso2.com>