Tweak color of Command Palette headings
Summary: The color was meant to match the background behind it, but it wasn't quite right.

Test Plan: Try the search `script:px/` in the Command Palette. If your screen (and eyes) are sharp enough, you should notice a very slight difference in the color of the heading compared to the content around it without the change; they should match with the change.
It was off by just one in the (r,g,b) tuple, so it's very subtle.
Light mode intentionally isn't using the same color, so that should still be noticeable.

Type of change: /kind bug

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>