[artifact manifests] Add `ArtifactManifest` type to represent a manifest file (#1096)
Summary: This PR adds `ArtifactManifest` which is a golang
representation of an artifact manifest file. In the future, the artifact
tracker will pull a manifest file from GCS, decode it into this type,
and then query for artifacts from `ArtifactManifest` directly. This PR
adds support for json serialization/deserialization for artifact
manifests. Internally, the artifacts in the manifest are stored in
version sorted order, such that newer versions are first. This will
enable faster version-ordered queries of manifests in the future.

Type of change: /kind cleanup

Test Plan: Added tests for `ArtifactManifest`
serialization/deserialization.