[artifact manifests] Allow querying `ArtifactManifest` for artifacts (#1104)
Summary: Adds `GetArtifact` and `ListArtifacts` methods to the
`ArtifactManifest`. `GetArtifact` returns an artifact with the given
name and version. `ListArtifacts` returns a list of artifacts of the
requested size, in version-sorted order with the newest versions first.
`ListArtifacts` accepts filters as well to eg filter out prereleases
from the list, or filter to only artifacts with a given artifact type
available.

Type of change: /kind cleanup

Test Plan: Added tests for `GetArtifact` and `ListArtifacts`, also
tested end-to-end as part of a change to artifact tracker.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>