Update dark color palette
Summary: Simplifies to fewer colors, but still very similar to the previous ones.
Footer and `nullish` strings in table details are the only ones that are easy to notice changed; they're more muted now.
The light theme and graph colors will get their own changes when design updates are ready for them.

Relevant Issues: #1174

Test Plan: Look at the live and admin views side-by-side with and without the change.
Light mode should be unaffected.
Graphs should be unaffected.
The footer and `nullish` fields in table details are a bit more muted.
The blue in table details is almost imperceptibly different.
A few of the neutral grays are almost imperceptibly different.

Type of change: /kind cleanup

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>