Upgrade openresty and switch to an alpine based image (#1178)
Summary: The debian based openresty image has a lot of open CVEs due to
all the default packages and libs included in the base distro. Switch to an
alpine based image effectively lets us close all of these issues.
The alpine image also includes sed which is used by our entrypoint
script but is missing bash, so we switch to using sh instead. Ensured 
that the entrypoint still behaves as expected.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Deployed a skaffold dev cloud. Ensured that the proxy service
still works and the UI loads and works as expected.

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>