[1 of 4] Add `UnitConnector` to Stirling, useful as an API to one source connector. (#1062)
Summary: We add class `UnitConnector` to wrap and provide an API to a
source connector. This wrapper provides the necessary logic to drive a
source connector in isolation (i.e. without using `stirling.cc`). It
will be used for test fixtures (e.g. in the socket tracer and perf
profiler) and for individual binaries (for the profiler, and for planned
work on record and replay binaries).

Type of change: /kind feature

Test Plan: This code is not used (yet). In the next set of PRs, we will
switch over test fixtures and stirling binaries to use the
`UnitConnector`. Continued passing tests will be the test plan as we
switch code over. See #1065.

---------

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>