[etcd] Create a multiarch compatible etcd image (#1229)
Summary: Upgrades to a modern version of etcd (3.5.8) in order to
support multiarch etcd. (Aside: the old version of etcd we were using
wasn't even an OCI compatible image, so we couldn't manually create a
multiarch image). The new version of etcd is now based on a distroless
image, so our startup/shutdown/healthcheck scripts wouldn't run. This PR
creates a new docker image based on Alpine that includes the etcd
binaries from the 3.5.8 etcd image. The new image packages the
startup/shutdown/healthcheck scripts into the image, instead of keeping
them in the kubernetes statefulset definition.

Relevant Issues: #893.

Type of change: /kind cleanup.

Test Plan: Along with the PR that updates the statefulset definition,
tested the new etcd on both x86 and ARM. Tested that I can make a change
to the yaml and redeploy, and the etcd cluster with successfully apply a
rolling update.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>