Upgrade prometheus library (#1253)
Summary: This upgrades the prometheus library being used to address
CVE-2019-3826. Note that prometheus versioning is a bit inconsistent and
v2.x releases were all removed and remapped to corresponding v0.x
releases for the module.
https://github.com/prometheus/prometheus#prometheus-code-base

The rest of the changes are automatic upgrades to transitive deps or
manual upgrades to make the various libraries work with each other
and changes to the code to fix API breakages.

Relevant Issues: CVE-2019-3826

Type of change: /kind cve,cleanup

Test Plan: All the existing unit tests pass. Will test a skaffold based
cloud and vizier.

---------

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>