Move graph widget controls into titlebar
Summary: TSIA. Affects network graphs, request graphs, and flamegraphs.
Also presses the edges of these graphs to the edge of their container to use the full available space.

Relevant Issues: #1174

Test Plan: Run `px/cluster` and `px/perf_flamegraph`.
The outermost borders of graphs in both should be pushed to the edges of their containers.
The extra buttons (including zoom buttons) for the request graph should be in the titlebar now, and still function.

Type of change: /kind feature

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>