[perf_tool] Exit early on errors in metric recorders (#1264)
Summary: Errors often occur in the metric recorders when a cluster goes
unhealthy or a query fails for other reasons. Previously, these errors
would only be surfaced after the experiment finished running. This PR
adds a channel to send the errors to, so that the errors are caught as
soon as they happen and the experiment can be abandoned earlier.

Type of change: /kind test-infra

Test Plan: Tested that sending an error to the error channel leads to
the experiment being aborted early.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>