Simplify `UnitConnector` by removing the "run window" concept. (#1263)
Summary: In the previous impl. of `UnitConnector`, we used the "run
window" concept borrowed from the main loop in Stirling proper. Since
the unit connector handles only one data source at a time, the run
window is not needed. In this patch, we remove the run window which
simplifies and cleans the code.

In addition, we tidy a few other features. We rename
`RefreshContextAndDeployUProbes` as `BlockAndDeployUProbes` to better
match what happens (i.e. because context refresh is not needed when
using the "everything local" context). Also, we add `SetClusterCIDR` to
support future usage in the socket tracer test suite.

Type of change: /kind cleanup

Test Plan: Tested locally with the next diffs that add to perf profiler
tests and convert the stirling profiler binary. The underlying patch was
developed with the effort to de-flake socket tracer tests.

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>