Remove update to BPF map not used in status quo tls tracing (#1298)
Summary: Remove update to BPF map not used in status quo tls tracing

In order to vet the new style of tls tracing, we introduced a mechanism
for detecting mismatched fds (#1161). This instrumented all of our tls
tracing when it was first developed. When the new method of tls tracing
was introduced, we removed the mismatched fd detection from the status
quo tls tracing (#1123), however, this BPF map update was missed in that
refactor (#1123).

Relevant Issues: #692

Type of change: /kind bug

Test Plan: Existing tests pass

Signed-off-by: Dom Del Nano <ddelnano@pixielabs.ai>