Add TLS probes to additional OpenSSL funcs to support tracing for Python 3.10 and later (#1338)
Summary: Add TLS probes to additional OpenSSL funcs to support tracing
for Python 3.10 and later

Python v3.10+ uses the `*_ex` version of the SSL read and write
functions. In this PR we expand our eBPF tracing to handle this variant.
In particular, we find the number of bytes read or written using the
function args. vs. the return value which was used for the non-ex
variants of SSL read & write.

By expanding our eBPF probes to handle the `*_ex` variants of ssl read &
write, we enable TLS tracing for Python 3.10+, but only when Python is
not compiled with `--enable-shared` (See #1347 for more details).

Additionally, we introduce a new test case using a Python 3.10 docker
image. This test case, which is the default image, is compiled with
`--enable-shared` and thus requires adding uprobes to
`libpython.<version>.so`. The test case has the effect of expanding
coverage for both `_ex` variants of ssl read & write and for Python
compiled with `--enable-shared`.

However, the scope of this PR was not intended to include a release to
support tracing Python binaries compiled with `--enable-shared`, rather,
that additional feature is planned for a later PR with more diligence
behind covering different Python versions. Therefore, we introduce a
compile time ifdef PL_BPF_TESTING to enable shared Python tracing for
testing (to cover both ex and shared) and to keep the shared Python
feature out of our releases for now.

Relevant Issues: #1113 #1347

Type of change: /kind feature

Test Plan: Added an `openssl_trace_bpf_test` for a Python 3.10
application and existing test coverage

Changelog Message:
```release-note
Add support for TLS tracing applications using Python 3.10 and later
```

---------

Signed-off-by: Dom Del Nano <ddelnano@pixielabs.ai>