Add k8ssandra demo support to CLI (#1367)
Summary: We want to add a K8ssandra demo to our list of demos. A few
changes needed to be made to ensure the demo can be deployed properly:
- We should check for cert-manager as a required dep. We decided not to
deploy this with the demo, since deploying 2 cert-managers can break
things. If we deploy cert-manager for the user, there's a bit of
bookkeeping we'd need to do if the user runs `px demo delete` to decide
whether or not we should also clean up cert-manager. Instead, we'll
leave this up to the user for the time-being.
- The demo deploys some CRDs. This can lead to race conditions if
deploying everything in the manifest at once. To handle this, added a
retry in the deploy block.

Relevant Issues: #680

Type of change: /kind cleanup

Test Plan: Push to dev demo manifest, `px demo deploy` with and without
cert-manager installed.

---------

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>