`perf_profiler.cc|h`: Implment better metrics for overrun of expected number of stack trace samples. (#1403)
Summary: We implement some additional metrics for case where the
profiler detects an overrun vs. the number of expected stack trace
samples. Although this overrun may not cause issues (because of
over-provisioning), we want to understand the magnitude and frequency of
the issue so that we can re-assess our provisioning.

Type of change: /kind feature

Test Plan: Existing tests. We know these metrics are reporting
infrequent overruns, we expect to see more granular data after these new
metrics roll out. Also, added some [verbose
logging](https://phab.corp.pixielabs.ai/P400) to capture both expected
values and actual values in the metrics. For this experiment, we lowered
the provisioned number of stack traces to ensure that the overflow
condition was triggered.

---------

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>