Reapply service accounts during operator update (#1515)
Summary: We used to filter out service accounts when reapplying the
Vizier YAMLs during an update. This is because infinitely-lived service
tokens would be created every single time as K8s secret, and lead to a
very long list of secrets. However, that has since been phased out in
K8s 1.24 and we want to make sure we properly handle cases where new
service accounts need to be created.
An alternative to this is to still filter out service accounts, but
check which ones exist already. However, that adds more network calls to
the process.

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Deploy operator with skaffold

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>