Fix dropdown selection in admin cluster detail view (#1558)
Summary: A wrong variable reference was causing "unknown cluster" to be
selected instead of what the user clicked.

Relevant Issues: N/A

Type of change: /kind bugfix

Test Plan: Visit `/admin/clusters`. Click one. In the dialog that comes
up, use the dropdown in the top left to choose another.
Now, it should reload the dialog with info for the new cluster
correctly.

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>