Let the Operator get the VizierID and send to Config Manager Service (#1670)
Summary: This PR lets the operator get the VizierID from the cluster's
secrets and then sends it to the config manager service for it to
ultimately set feature flags.

Relevant Issues: Fixes #1632

Type of change: /kind bug

Test Plan: Skaffold the cloud changes on the testing cluster and then
skaffold the operator changes on a dev cluster. Manually trigger an
update on the operator side to send the VizierID to the config manager
service. Check the config manager service's logs to see that the new
logic is being executed as expected to handle the VizierID.

---------

Signed-off-by: Kartik Pattaswamy <kpattaswamy@pixielabs.ai>