[Fix CQL Stitcher 2/4] Update StitchFrames to use new interface (#1715)
Summary: Fixes the head of line blocking issue discussed in #1375 by
using a map of stream_ids to deque of frames, which we populate in the
`conn_tracker`. CQL stitcher tests are updated to reflect the bug fix
and new interface.

Related issues: #1375

Type of change: /kind bug

Test Plan: Tested all existing targets and updated the unittests for the
CQL stitcher. `Note`: this PR relies on changes introduced in #1689.

Signed-off-by: Benjamin Kilimnik <bkilimnik@pixielabs.ai>