[Fix CQL Sticher 3/4] Populate map of streams to frames during parsing (#1716)
Summary: Modifies all protocol parsers to use a map of streams to deques
by default. Protocols which do not have a notion of streams are encoded
as single keys in a map. This completes the CQL stitcher fix and should
simplify stitching frames for protocols with streams.

The final PR in this sequence #1732 populates a map of streamIDs to
deque of frames in ParseFramesLoop instead of ParseFrames. This should
provide a small efficiency boost, as we won't have to loop over the
frames twice.

Related issues: Closes #1375

Type of change: /kind bug

Test Plan: Updated parsing tests to use new interface. `Note`: this PR
relies on changes introduced in #1689 and #1715

---------

Signed-off-by: Benjamin Kilimnik <bkilimnik@pixielabs.ai>