[standalone pem] Fix execution stats. (#1737)
Summary: We update the code to populate the execution stats data using
`mutable` so that the data is written into message that will be sent on
the wire. Previously, our code created a copy of the execution stats and
updated that. The copy left the upstream message empty and consequently
nothing was sent on the wire.

Type of change: /kind bug fix.

Test Plan: We discovered this bug when testing with the px go api.
Standalone pem would work when using streaming, but the client api
program would fail for batched. The cause of the failure was an
unidentified message data type, i.e. a completely empty execution stats
data field. We successfully re-tested the batched query with a
standalone pem and the go px api.

---------

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>