Update kernel builder to work with kernels 6.6 and later. Add 6.8.x to kernel build list (#1992)
Summary: Update kernel builder to work with kernels 6.6 and later. Add
6.8.x to kernel build list

Sicne Ubuntu's current LTS release runs on a 6.8.x kernel, I wanted to
keep our newest kernel version in line with that. If/when this is
approved, I'll proceed with running the
[upload](https://github.com/pixie-io/pixie/blob/33244cd28d79058792a6cc06fe2637f491cd51a7/tools/docker/Makefile#L288)
(side stepping the GCS part) and then add this to the
[all_kernel_versions](https://github.com/pixie-io/pixie/blob/33244cd28d79058792a6cc06fe2637f491cd51a7/ci/github/matrix.sh#L31)
ci matrix.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: Ran the makefile target for an older and newer kernel to
verify it works now

---------

Signed-off-by: Dom Del Nano <ddelnano@gmail.com>